Go back to using __vector unsigned long long for POWER8 and above

+++ TestPrograms/test_ppc_aes.cxx
@@ -8,13 +8,13 @@ int main(int argc, char* argv[])
 	x=__vncipher(x,x);
 	x=__vncipherlast(x,x);
 #elif defined(__clang__)
-	__vector unsigned long x = {1,2};
+	__vector unsigned long long x = {1,2};
 	x=__builtin_altivec_crypto_vcipher(x,x);
 	x=__builtin_altivec_crypto_vcipherlast(x,x);
 	x=__builtin_altivec_crypto_vncipher(x,x);
 	x=__builtin_altivec_crypto_vncipherlast(x,x);
 #elif defined(__GNUC__)
-	__vector unsigned long x = {1,2};
+	__vector unsigned long long x = {1,2};
 	x=__builtin_crypto_vcipher(x,x);
 	x=__builtin_crypto_vcipherlast(x,x);
 	x=__builtin_crypto_vncipher(x,x);
